+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/member/MemberService.java
@@ -2,7 +2,7 @@ package com.startingblue.fourtooncookie.member;
 
 import com.startingblue.fourtooncookie.member.domain.Member;
 import com.startingblue.fourtooncookie.member.domain.Role;
-import com.startingblue.fourtooncookie.member.dto.request.MemberSaveRequest;
+import com.startingblue.fourtooncookie.member.dto.MemberSaveRequest;
 import com.startingblue.fourtooncookie.member.exception.MemberDuplicateException;
 import com.startingblue.fourtooncookie.member.exception.MemberNotFoundException;
 import com.startingblue.fourtooncookie.member.service.MemberDiaryService;
